restructure code by splitting api.py into two parts adding Nordfyns to same API as Viborg, needs tests...
